APIMF-612: add custom emitter in SHACL validation for scalars

+++ amf-webapi/shared/src/main/scala/amf/plugins/domain/shapes/metamodel/CommonShapeFields.scala
@@ -22,7 +22,7 @@ trait CommonShapeFields {
 
   val Format = Field(Str, Shapes + "format")
 
-  val MultipleOf = Field(Int, Shapes + "multipleOf")
+  val MultipleOf = Field(Double, Shapes + "multipleOf")
 
   val commonOASFields =
     List(Pattern, MinLength, MaxLength, Minimum, Maximum, ExclusiveMinimum, ExclusiveMaximum, Format, MultipleOf)
